+++ tests/testthat/test-nn.R
@@ -360,3 +360,15 @@ test_that("nn_module_list names", {
   )
    
 })
+
+test_that("deduplicate duplicated parameters", {
+  m <- nn_module(
+    initialize = function(x) {
+      x <- nn_linear(10, 10)
+      self$x <- x
+      self$y <- x
+    }
+  )
+  expect_length(m()$parameters, 2)
+  expect_named(m()$parameters, c("x.weight", "x.bias"))
+})